Add Dynamic sorting for Taxonomies
Moved sorting to Model so Taxonomy can be dynamically sorted on recalculate

Closes #4968
